copp::robot

Trait RobotBasic

pub trait RobotBasic {
    // Required method
    fn dim(&self) -> usize;
}

Minimal robot metadata required by the planner.

A usize variable can serve as a trivial RobotBasic implementation representing the robot dimension, but users can also implement this trait for their own robot models.

Required Methods

fn dim(&self) -> usize

Return robot dimension / DoF.